move HorizontalSelector out of its own directory
